+++ src/lib.rs
@@ -41,6 +41,8 @@
 #![feature(coroutines, proc_macro_hygiene, stmt_expr_attributes)]
 #![feature(binary_heap_into_iter_sorted)]
 #![feature(let_chains)]
+#![feature(context_ext)]
+#![feature(local_waker)]
 
 /// Protocol codec
 mod codec;
@@ -75,7 +77,7 @@ pub mod server;
 /// Raknet client
 pub mod client;
 
-/// Connection options
+/// Connection optional settings
 pub mod opts;
 
 #[cfg(feature = "micro-bench")]